Fix detached menu when TEdit passed notitle window
Refactor to always create a window and pass it to TEdit, which makes the
code clearer and more uniform. Also remove an unnecessary binding.

If TEdit is passed a window with no title the menu window may end up
detached from TEdit if the title is set after opening TEdit. So set the
window title BEFORE calling TEdit.